requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
importlib-metadata==4.5.0
iniconfig==1.1.1
jsonpath==0.82
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyhamcrest==2.0.2
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
